Made the debug filter dropdown its own widget
